+++ src/hooks/recruiting/index.ts
@@ -1,0 +1,4 @@
+// src/hooks/recruiting/index.ts
+
+export * from "./usePipeline";
+export * from "./useRecruitingStats";